Implemented registration tab in React Native app with Input component, checkboxes, buttons, and text
Created a new component called "Input" to encapsulate the logic and rendering of text labels and TextInput fields.
Integrated the Input component into the registration tab of the React Native app.
Implemented three instances of the Input component for capturing user input for name, email, and password.
Added two checkboxes to allow users to select optional features or agreements.
Included two buttons, one for registering with user-provided data and another for registering with Google.

Co-authored-by: DerekAyala <derek.ayala12345@hotmail.com>
Co-authored-by: arturovaldez019 <arturovaldez019@gmail.com>