Create a new view for the login screen, components for the buttons and added navigation between login and register screens.
Created a new login.tsx file and added the code for the login screen.
This file used the input and button components created before.
Also created a new buttons.tsx file and added the code for the buttons.
This buttons get a label prop to set the label of the button and be able to reuse the component.
Added navigation between login and register screens using react navigation.
For this we added a new stack navigator in App.tsx and added screens for login and register, this allow us to call any of this screens from any other screen.

Co-authored-by: DerekAyala <derek.ayala12345@hotmail.com>
Co-authored-by: arturovaldez019 <arturovaldez019@gmail.com>