Create a new folder for the styles and separate the styles for each component into their own files.
Created a new folder called styles and moved the styles for each component into their own files.
Each file is named after the component it is styling.
Also created a library file for the colors used in the app, to be easily accessible.
To make this changes, I had to change the imports in the components and screens to reflect the new file structure.

Co-authored-by: DerekAyala <derek.ayala12345@hotmail.com>
Co-authored-by: arturovaldez019 <arturovaldez019@gmail.com>