Modify the login and register screens to use the new components
In this commit we create a new folder called hooks, and inside it we create a new file called useTogglePasswordVisibility.ts. This file contains a custom hook that we will use to toggle the visibility of the password in the login and register screens.
Fixed the eslint errors in the AppVuelos/.eslintrc.js file.
Modify login and register screens with the new input functionality.
Added checkbox functionality to the register screen.
Modify the ButtonStyle and InputStyles files with new styles.

Co-authored-by: DerekAyala <derek.ayala12345@hotmail.com>
Co-authored-by: arturovaldez019 <arturovaldez019@gmail.com>