Created validations for login and register forms
In this commit we created the validations for the login and register forms.
Created a validations for the buttons to be active only when the form is valid and all the fields are filled.
Created a validation for the email field to be a valid email.
Created a validation for the password field to be at least 8 characters long.
Unified the styles of the login and register forms in only one new file.

Co-authored-by: DerekAyala <derek.ayala12345@hotmail.com>
Co-authored-by: arturovaldez019 <arturovaldez019@gmail.com>