Add navigation to a test screen when the user is logged in
Added a test screen to test the navigation when the user is logged in or the register is valid.
Updated the login and register screens to navigate to the test screen when the user is logged in or the register is valid.
Updated App.tsx to fix navigate error.
Added new error message to the login and register screens.
On register screen added a new error message when the user is already registered and the button now stays disabled until the password match the regex pattern.
For the login screen added a new error message when the user is not registered or the password is incorrect.

Co-authored-by: DerekAyala <derek.ayala12345@hotmail.com>
Co-authored-by: arturovaldez019 <arturovaldez019@gmail.com>